refactoring

+++ scripts/compile.sh
@@ -1,7 +1,5 @@
 #!/bin/bash
 
-CLIENT_SOURCE="./cmd/client/main.go"
-CLIENT_TARGET="./build/client"
 SERVER_SOURCE="./cmd/server/main.go"
 SERVER_TARGET="./build/server"
 
@@ -9,8 +7,6 @@ KEY=$(openssl rand -base64 16)
 VAL=$(openssl rand -base64 16)
 TOKEN=$(echo ${VAL} | openssl dgst -sha256 -hmac ${KEY} -binary | openssl enc -base64 -A)
 echo ${TOKEN} > ./security/token
-
-VERSION_INITIAL="00.01.00"
 
 # Create shortcuts for functions and static files
 if [ ! -e "./functions" ] || [ ! -e "./static" ]; then
@@ -20,19 +16,14 @@ fi
 
 # Rename the module if needed
 MODULE_OLD=$(cat go.mod | head -1 | cut -f2 -d' ')
-rm -rf go.* && go mod init &>/dev/null
-MODULE_NEW=$(cat go.mod | head -1 | cut -f2 -d' ')
-if [ ! "${MODULE_OLD}" == "${MODULE_NEW}" ]; then
+BASENAME_OLD=$(basename ${MODULE_OLD})
+BASENAME_NEW=$(basename ${PWD})
+if [ ! "${BASENAME_OLD}" == "${BASENAME_NEW}" ]; then
+    rm -rf go.* && go mod init &>/dev/null
+    MODULE_NEW=$(cat go.mod | head -1 | cut -f2 -d' ')
     for FILE in $(find . -name "*.go"); do
         sed -i 's|'${MODULE_OLD}'|'${MODULE_NEW}'|g' $FILE
     done
-fi
-# Initialize git
-if [ ! -d ".git" ]; then
-    git init
-    git add .
-    git commit -m "initial commit" .
-    git tag ${VERSION_INITIAL}
 fi
 
 mkdir -p "./build"
@@ -49,7 +40,9 @@ minify -r -b -o ./static/bundle.css ./static/*.css &>/dev/null
 # Copy the bundle
 rm -f ./bundle/*
 cp -f ./static/*.htm* ./bundle/ &>/dev/null
+cp -f ./static/*.ico ./bundle/ &>/dev/null
 cp -f ./static/*.json ./bundle/ &>/dev/null
+cp -f ./static/*.png ./bundle/ &>/dev/null
 cp -f ./static/*.svg ./bundle/ &>/dev/null
 cp -f ./static/bundle.* ./bundle/ &>/dev/null
 for FILE in $(find ./bundle/ -name "*.*"); do
@@ -61,11 +54,12 @@ export CGO_ENABLED=0
 FAASIFY_TOKEN=${TOKEN} go build -ldflags "\
     -s -w" \
     -o ${SERVER_TARGET} ${SERVER_SOURCE}
+
+export GOARCH=amd64
 FAASIFY_TOKEN=${TOKEN} go build -ldflags "\
     -s -w" \
-    -o ${CLIENT_TARGET} ${CLIENT_SOURCE}
+    -o ${SERVER_TARGET}_amd64 ${SERVER_SOURCE}
 
 # Minify binary
-upx ${CLIENT_TARGET}
 upx ${SERVER_TARGET}
-
+upx ${SERVER_TARGET}_amd64
